# Build validator copy cache on write