This was just fun to play with: Swap cache map on write. 😬 

Also tried a waitgroup, but that has worse performance than the r/w lock.
 
Idea for: https://github.com/bufbuild/protovalidate-go/issues/29

```
goos: darwin
goarch: arm64
pkg: github.com/bufbuild/protovalidate-go
                               │  current.txt  │               new.txt                │
                               │    sec/op     │    sec/op     vs base                │
Validator/ColdStart-8             210.9µ ± ∞ ¹   211.8µ ± ∞ ¹       ~ (p=1.000 n=1) ²
Validator/Lazy/Valid-8            136.1n ± ∞ ¹   104.3n ± ∞ ¹       ~ (p=1.000 n=1) ²
Validator/Lazy/Invalid-8          391.9n ± ∞ ¹   386.9n ± ∞ ¹       ~ (p=1.000 n=1) ²
Validator/Lazy/FailFast-8        136.90n ± ∞ ¹   97.42n ± ∞ ¹       ~ (p=1.000 n=1) ²
Validator/PreWarmed/Valid-8       101.1n ± ∞ ¹   102.8n ± ∞ ¹       ~ (p=1.000 n=1) ²
Validator/PreWarmed/Invalid-8     382.5n ± ∞ ¹   385.7n ± ∞ ¹       ~ (p=1.000 n=1) ²
Validator/PreWarmed/FailFast-8    93.68n ± ∞ ¹   95.46n ± ∞ ¹       ~ (p=1.000 n=1) ²
geomean                           476.5n         439.2n        -7.84%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

                               │  current.txt  │                new.txt                │
                               │     B/op      │     B/op       vs base                │
Validator/ColdStart-8            421.4Ki ± ∞ ¹   421.4Ki ± ∞ ¹       ~ (p=1.000 n=1) ²
Validator/Lazy/Valid-8             0.000 ± ∞ ¹     0.000 ± ∞ ¹       ~ (p=1.000 n=1) ³
Validator/Lazy/Invalid-8           649.0 ± ∞ ¹     649.0 ± ∞ ¹       ~ (p=1.000 n=1) ³
Validator/Lazy/FailFast-8          168.0 ± ∞ ¹     168.0 ± ∞ ¹       ~ (p=1.000 n=1) ³
Validator/PreWarmed/Valid-8        0.000 ± ∞ ¹     0.000 ± ∞ ¹       ~ (p=1.000 n=1) ³
Validator/PreWarmed/Invalid-8      648.0 ± ∞ ¹     649.0 ± ∞ ¹       ~ (p=1.000 n=1) ²
Validator/PreWarmed/FailFast-8     168.0 ± ∞ ¹     168.0 ± ∞ ¹       ~ (p=1.000 n=1) ³
geomean                                      ⁴                  +0.02%               ⁴
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05
³ all samples are equal
⁴ summaries must be >0 to compute geomean

                               │ current.txt  │               new.txt                │
                               │  allocs/op   │  allocs/op    vs base                │
Validator/ColdStart-8            5.798k ± ∞ ¹   5.800k ± ∞ ¹       ~ (p=1.000 n=1) ²
Validator/Lazy/Valid-8            0.000 ± ∞ ¹    0.000 ± ∞ ¹       ~ (p=1.000 n=1) ³
Validator/Lazy/Invalid-8          15.00 ± ∞ ¹    15.00 ± ∞ ¹       ~ (p=1.000 n=1) ³
Validator/Lazy/FailFast-8         3.000 ± ∞ ¹    3.000 ± ∞ ¹       ~ (p=1.000 n=1) ³
Validator/PreWarmed/Valid-8       0.000 ± ∞ ¹    0.000 ± ∞ ¹       ~ (p=1.000 n=1) ³
Validator/PreWarmed/Invalid-8     15.00 ± ∞ ¹    15.00 ± ∞ ¹       ~ (p=1.000 n=1) ³
Validator/PreWarmed/FailFast-8    3.000 ± ∞ ¹    3.000 ± ∞ ¹       ~ (p=1.000 n=1) ³
geomean                                     ⁴                 +0.00%               ⁴
```